requirements.txt
Django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
python-decouple==3.4
dj-database-url==0.5.0
psycopg2==2.8.6
django-filter==2.4.0
django-rest-passwordreset==1.1.0
whitenoise==5.2.0
coreapi==2.3.3
djangorestframework-gis==0.17
django-sortedm2m==3.0.2
geopy==2.1.0